フォーマット選択画面のレイアウトを作成 ファイル名をリネーム
